Update all test base classes to use base.TestCase
We have a centrally defined test case base class that handles a set of
things like log capture, mocking time.sleep and setting up
requests-mock.

We also had a split base test case, with some things using TestCase and
some using RequestsMockTestCase. This was a holdover from the transition
to requests-mock. We are finally at the point where we don't need the
split, so merge TestCase and RequestsMockTest case.

Then, update all of the tests to use the new combined base class.

Also, replace a use of unittest.skipTest with self.skipTest from the
base class.

Change-Id: I2cc3e201a5241262e5d102d3de8423c4fb2a8c4a