Add all_projects as a preferred alias for all_tenants
Now that we support both server-side and client-side names for
parameters, we can add an all_projects alias. We already use project_id
in resources instead of tenant_id.

Put in logic to only pass all_projects along to the underlying list if
it's True. That way we can easily use the interface with True/False and
only send query parameters when we need to. Pehaps we could enhance
QueryParameters to understand default values of parameters and that they
don't need to be sent?

Change-Id: I66117ab4c82d30ae3700e17f7598b3d3fbbac60d