Fix issue with missing url parameters
OSC uses these. project_id and tenant_id for neutron are interchangable.
If we list them in the QueryParameters list as project_id='tenant_id'
then a user can specify either. We'd done this for all of the *other*
Resources, but for some reason Network was left out.

ip_version is a valid query parameter for network_ip_availability, and
we even have it documented, but it's not in the list either.

We should have caught this before the last release - but actually only
just now got osc-functional-tips jobs turned back on. Sigh.

Change-Id: I1037dac37b15edbd2b8a17d7ed4ba639cf8e960b