Run normalize_keys on config for session codepath
When we pass an authenticated session to Connection, we use the
from_session codepath. That doesn't run through loader, and as a result
the kwargs dict doesn't get passed through normalize_keys. That, in
turn, means that int _api_version values are not turned in to strings,
which in turn breaks a subscripting attempt in service_description.

Run normalize_keys on the config dict and add some tests for this.

Change-Id: I1c745c45512cb5d10cad36a3e2713a1dec9d494f