Remove the need for OpenStackConfig in CloudRegion
There are a few things in CloudRegion that require calling back out to
the creating OpenStackConfig object. However, now that creating a
CloudRegion directly is a thing, those calls have become more awkward.

Shift the cache settings and methods to CloudRegion along with the
extra_config logic. Move two of the utility methods in loader into a
_util file so that both loader and cloud_region can import it. This
renames get_cache_expiration to get_cache_expirations, but I'm fairly
confident that the only consumer of get_cache_expirations here is
openstack.cloud.

Leave passing in the OpenStackConfig for now, because osc uses it.
However, if we can get osc shifted to using get_password_callback, we
can get rid of it.

Change-Id: Ia6ed0f00bfc2483bd09169811198cdf1a0ab2f15