Shift tag resource definition to TagMixin
The TagMixin class defines things needed for resources that support
tags. Currently the use of this involves a copy-pasta'd chunk (complete
with mis-spelling) to each class that's using the mixin. Move the tags
resource to the Mixin and remove it from the indivdual classes so that
we don't have to copy/paste the chunk.

Also, fix the mis-spelling.

Change-Id: If4942cb0e58e0fb51125c53cf1cfae9d65c57cd0