Allow not resolving outputs on get stacks
Depending on the stack complexity, resolving outputs of a Heat stack can
be fairly expensive. The API has a parameter to disable output
resolution: this adds this parameter to the get_stack call, and use it
in places where outputs are superfluous.

This is a copy of shade commit b1bc65c599812a02f6861c1d80c31c3b9be894da.

Change-Id: I932e42fe10af7de7d3a08182328ff2c8c70e8f9e