Prepare for os-client-config wrapper
In order to remove the second copy of the code that is in
os-client-config, we need to be able to set a few class/module names to
allow openstacksdk methods to return the appropriate os-client-config
wrapper subclasses when called via os-client-config.

Also remove a function we had intended to remove back in 2015 that
should not be used by anyone.

Change-Id: Ie279c543fb2142c101f03dcdd45314bbc098ebc4