Temporarily disable osc-functional-devstack-tips
There is a deeper issue related to versions and sibling installation in
the period between a release and the first release in the next cycle on
the master branch. A fix is in progress for that, but it will take a few
more days to get it in place.

In the meantime, the workaround is actually to just cut a release of
openstacksdk. The floating ip patch (next) is a pretty important bug, so
disable the osc-functional-devstack-tips job for a second so that we can
land the floating ip fix. Then we can cut a new release and re-enable
the job.

A revert of this patch follows.

Change-Id: Idec38e3017db57ea56b3bce860a261abee7076d7