Fix devstack tests
With the change in devstack from neutron-legacy to the new neutron lib
setup, we need to tweak things a bit.

Include neutron devstack plugin, that would have been needed for some
time in order to make the neutron-qos service work.

Depend on octavia change to make its devstack plugin compatible.

Depends-On: https://review.openstack.org/544281
Change-Id: I023d76c56d9b72e390bf2c1d39ddc27584d5c52a