Fix DNS Recordset CRUD
If a zone name is supplied to a recordset function
we should lookup the ID to use that in the URL,
as the name in the URL will return a 404.

The `name` parameter for a recordset needs to be the full FQDN
including the zone part. Fix the tests accordingly.

Looking up recordsets via name does not work tests for that behaviour
have been removed.

Also harmonised the naming of vars in the recordset
section to make this clearer.

Also add designate to the services deployed for functional tests

These issues with the DNS related cloud functions went undiscovered
for some time because even though there are tests for it,
those get skipped when designate isn't deployed.

Change-Id: I60d6400631f31cba89a6d07382c380006bd2ded2
Signed-off-by: Graham Hayes <gr@ham.ie>
Co-Authored-By: Jens Harbott <j.harbott@x-ion.de>