Fix F405 errors
F405 is a bit more aggressive about not being able to tell if a
symbol is valid due to * imports. Just go ahead and fix it since *
imports are, in fact, hard to deal with.

As a result of fixing this, there are two actual bugs that were hidden
that the check found, so the fix seems worthwhile.

Change-Id: If76b0943e16e0c7264b75829838c4a715342ab9d