Avoid globally modifying yaml library
The heat utils set up the yaml parser to always return unicode strings,
but are currently doing it in a way that infects all uses of the
library.

Make a subclass of the loader and run add_constructor on that so that
the new constructor is confined to the specific loader.

Change-Id: I49e97b5e2ae9b7862915ff83263718cf6cad32b8
Story: 2002040