allow passing ansible variables to ansible tests
Provide possibility to pass ansible variables through environment (i.e.
ANSIBLE_VAR_FLAVOR for setting `flavor` variable in ansible playbooks)
to the ansible functional tests. This can be used i.e. for running tests
in other than devstack cloud, where images, flavors, networks are not
matching devstack.

This is an initial change in a series of changes to enable executing
ansible tests in other clouds as a validation.

Change-Id: I40bc16241345ec4f39481f46dc0ad3a3075201fe