Import rate limiting TaskManager from nodepool
There is logic in the caching and batching layer in shade that is in
service of the rate limiting TaskManager from nodepool- but the logic in
that TaskManager is nowhere to be found in the openstacksdk codebase.
This leads people to want to make changes without the whole picture.

Pull in the code from nodepool so that we can have a full discussion
about things like caching and retries ... and potentially even add some
tests that run things under it.

Change-Id: I0fcefb78cf66a60a4c08de8d39d44222a56a8725