Move clustering additions from shade directories
In the cherry-pick from shade, the file locations weren't updated.

Also, several of the tests didn't work, which we didn't catch because
they weren't run when we gated the original change. Fix the tests.

Finally, the functional tests don't work, so disable them. The followup
patch will re-enable them so we can work through fixing them.

Change-Id: I4cf8639aaacedf64dcbd3bca1467a0d02d88350e