Support for microversions in baremetal resources
The baremetal resources are updated with a _max_microversion reflecting
the currently supported set of fields. Doing that would make create()
call create nodes in the "enroll" state, which would be a breaking change.
Thus, the create() call is updated to handle provision_state being set
to enroll, manageable or available, with available being the default.

Change-Id: If46e339070514bcd34bc3ba336f0cfb5556a5cea