Update create_object to handled chunked data
We try to calculate len for data, which breaks with chunked uploads, but
we only use that with filename. Move the code so that we don't try to
calculate length of things where we don't need to.

We still need to figure out streaming uploads and large objects.

This changes several unit tests due to the re-ordering of when
get_object_segment_size gets called. It's not necessary to call it for
the data path, so it was moved to later, which puts it after the
container create. For data paths this results in one less call. For file
paths, the calls are the same but the /info call happens later.

Change-Id: I841d7049ff2f05e00a31fecbc27dda6a0007be16