Pass microversion info through from Profile
Even though Profile is going away, senlinclient is still using it.
There is a patch up to fix that, but in the meantime, let's fix it for
their issue. api_version in Profile is the variable for setting a
default microversion, but the translation method didn't set it.

(cherry-picked from commit 6464422a2b911499556d4b82b44532f6d569857b)

Change-Id: Id8ed4a6c5d5a26a96944cf7f9dca6d3272784dac
Story: 2003146