Don't wait for task in submit_task
The exception shifting performed by the Task class is meant to
protect the TaskManager run method from encountering any exceptions
done in the course of running a task.  This is especially important
for subclasses of TaskManager which implement alternate run
strategies.  The exception is shifted so that it is raised by the
Task.wait method.  In a multi-thread environment, it is the caller
of the submit_task method which should receive the exception, and
therefore it is submit_task which should call Task.wait.  Currently
it is the _run_task method which calls Task.wait, which means the
TaskManager itself receives the exception.

Change-Id: I3a6e61601164811fdd255ae54470011768c99a7d