Rename Resource get and update to not clash with dict
The next change is going to make Resource a subclass of dict so that we
can return Resource objects in the shade layer. Unfortunately, the
Resource base class has two methods, get and update, that conflict with
standard dict method names. To keep things reviewable, just change the
method names.

Most consumers should not be using either of these methods directly.
They are mostly there for lower-level things to use. However, they COULD
be using them, so it's important to note that this is a breaking change.

Change-Id: I2fedeea6e405dcbd333482c1964173ade98ca04d