Add computed attribute type and location to base resource
The shade objects all have a 'location' attribute which includes
information about the cloud, domain and project an object is from.

Add this to the base Resource object so that Resource objects start to
match shade objects. Also, add a new type of attribute, 'computed' to
set location to since location is not necessarily information from the
remote payload.

Also add an attribute parameter "coerce_to_default" which can be set to
tell the SDK to force the use of the default value if the attribute
would have been None or missing.

Change-Id: I315bc863bc87a17fb6f6a3e265a46b9e0acd41ed