Make resource a dict subclass usable by shade layer
Users of shade/openstack.cloud expect munch.Munch objects to be returned
by all the methods. These objects are basically dicts that allow
object-notation access.

In an attempt to minimize the size of the mapping layer between shade
and sdk code, make the Resource objects behave like dicts as well as
like objects. We want the attributes exposed as dictionary attributes
to be the data model defined with the Component descriptors, so we wind
up needing to override the munch methods anyway - at which point munch
itself ceases being valuable.

he overall idea here is allowing the shade-layer API to return SDK
Resource objects directly without transforming them with to_dict first.
If we can make that work, it means that a user could use shade and sdk
methods interchangably, being able to do something like:

  # Create server using shade layer
  server = conn.create_server('my-server', auto_ip)
  # reboot server using sdk
  conn.compute.servers.reboot(server, 'HARD')

There is a REALLY ugly hack in here (with a comment) that's needed to
make json.dumps(server) work. It's not a hack we should actually use,
but my brain is too tired right now to figure out how to make it
actually work.

There is a latent bug in image. A TODO has been left.

Co-Authored-By: Rosario Di Somma <rosario.disomma@dreamhost.com>
Change-Id: Ib48f0bfa74231bf3171a798a179f6606177c95b0