Fix to_dict recursion issues with circular aliases
Resources that have two different attributes that are aliases of each
other cause to_dict to go into infinite recursion.

Because of how the attributes and descriptors work, set the first one to
None if it's not found, so that when the second one is tried it's either
returned or when it goes to try its own alias (the first value) it'll
hit the None and exit the recursion.

Change-Id: I5aff07311967c0150d74d1f834516e8e92d8a73b