baremetal: support newer microversions in {get,update,patch}_machine
This change switches get_machine and update_machine to use the baremetal
proxy. patch_machine still uses the session directly, but detects
a microversion by calling into Node._get_microversion_for.

API 1.1 (early Kilo) and older return None instead of "available".
This change forces the value to be available in this case. While
potentially breaking for the non-shade part, it improves consistency
in the SDK.

Change-Id: Iff682b9f307d11c8bb4f133e37e6d1cb0811edfd