Fix list_recordsets to deal with top-level key
During use of the os_recordset Ansible module, it was found that from
the change from shade to openstacksdk and newer versions of Ansible
and Designate that the module lost idempotence and the ability to
create DNS records using the filter in function "search_recordsets"
in openstack/cloud/openstackcloud.py

search_recordsets passes expected output to _utils._filter_list, but
this can not be done since new versions of Designate nest the output
in an array called "recordsets". If that is stripped out, then the
filter works and output is displayed.

Change-Id: If5e85d0b5b1a86ba6d05508a5615699aba1de78a
Story: #2003607