Handle empty values in regions
I had a cloud config file that looked like

 clouds:
   mycloud:
     regions:
       - name: region
     auth:
      ...

and the client became rather unhelpful:

 $ OS_CLIENT_CONFIG_FILE=./clouds.yaml openstack --os-cloud=mycloud ...
 'values'
 $ # huh?

Using '--debug' did give the traceback, but I still had to dig around
in the code to figure it out.

Anyway, it turns out that if a "regions:" entry is a dictionary, it is
assumed it has a "values:" key.  It seems like there's nothing really
wrong with just defaulting to a blank "values" in this case and
allowing it to be a valid config.  With this the above config was
handled correctly.

Add some validation on the regions entries to ensure it has a 'name'
key, and only the keys 'name' and 'values' while we're here.

Tests are added to cover both cases

Change-Id: I503b40a5d93d384d7d86268194245dc3e8d75744