Explicitly set logging levels for external libs
Previously logging levels were only set explicitly if we wanted debug
level. The problem with this is if you are including the sdk in a system
that sets a global debug logging level then you cannot set a less
verbose logging level in the sdk via enable_logging.

To fix this explicitly set the logging levels in enable_logging which
allows you to override any global application logging levels. We do this
for keystoneauth, urllib3, and stevedore.

Change-Id: I6ac281023b9fd8ccd89d7540169e591b4ea1a470