Clean up python3 test and split networking into a job
We want to run with python3 always with python2 being the exception.

Update the functional tests to be python3 by default with a
python2 job to check compat.

Remove python3.5 jobs, since that's covered by python3.6 and the
differences between 3.5 and 3.6 are covered by also having 2.7 in the
gate.

Similarly, make the tips jobs python 3.6 only because they are jobs that
aim to ensure interactions between openstacksdk and keystoneauth and
os-client-config and shade are correct at the unit test level. The
individual unit tests should catch actual python 2.7 issues.

Remove cover jobs because we historically haven't cared about the output
and they take up test nodes. In general it's not that big of a deal, but
we tend to be patch-heavy around here, so it feels extravagant to run a
job we don't actually care about the output of.

Split advanced networking services into a job

We're hitting timeouts like crazy- maybe we're just running low on
resources. Split octavia and designate into their own job. Turn
swift off in that job.

Skip sdk layer dns functional tests when designate is not present. We
should refactor these to deal with both cases.

Change-Id: Ica5a47cc200c8abff1d20af9883c5192fcbb95e1