Add support for per-service rate limits
Rate limits on the server-side are per-service, but the rate limit in
the TaskManager is a single rate. Add support for a dict of rate limits,
keyed by service-type.

The primary user interface should be passing rate to the Connection
constructor. That takes calls-per-second for both scalar and dict
versions of its rate parameter.

Depends-On: https://review.openstack.org/612168/
Change-Id: If0ff77b43adc1f6f0bb1e7c08908930a95508b31