Restore timeout_scaling_factor
In a previous refactor, we lost the part where we actually multiple
TIMEOUT_SCALING_FACTOR to the timeout. Whoops.

Also, turn off keepalive in the underlying session. For devstack
functional tests, we're seeing a lot of "Resetting dropped connection"
so I'm thinking we should just disable that.

Also, increase the individual test timeout, and bump the scaling factor
on trunk tests, which always seem to be timing out.

Change-Id: Ie5b3eb319fed10cabfb2e818f3fcfd52b6fd7b14