Rearrange shade image code
In anticipation of moving the image creation logic down into the sdk
image proxy class which has natural v1 and v2 classes, rearrange the
branching logic to start with v1/v2 then go to put/task. This will let
us do that reorg in place so that when we move things into proxy classes
it'll be more of a direct copy/paste.

There is one test update. In doing the code shuffle, there was a
duplicate unneeded call to get_image found in the task codepath. That
call was removed. However, to compensate, putting the wait in
create_image itself needlessly calls get_image after it has been done in
the task workflow. The next patch should be able to remove the second
get_image call.

Change-Id: Id76d1750e5af46b39d08428e0c6d9d8b30a43e5e