Revert the Proxy metaclass
Sometimes one can be too clever. The win of generating methods here
doesn't really outweigh the complexity or the increased difficulty
understanding the codebase. Additionally, we run in to an EXCELLENTLY
obtuse error:

  TypeError: metaclass conflict: the metaclass of a derived class must be
  a (non-strict) subclass of the metaclasses of all its bases

when trying to make an abstract base class for the image proxy base
class.

Just get rid of the metaclass and restore the (very few) generated
methods.

Change-Id: Ib53d7b29526a734f1dcf4088bf156e9a29746f5b