Support v4-fixed-ip and v6-fixed-ip in create_server
novaclient supports v4-fixed-ip and v6-fixed-ip as aliases for fixed_ip.
As shade used to use novaclient and passed this content through the
switch to REST resulted in a behavior regression for shade users.

openstackclient also exposes these for the --nics parameter, so it's
important that we also support it as we look forward to switching OSC to
sdk.

Change-Id: I3da91b6f18528b515a45d868e8c3faf824ade5a0