Make delete_unattached_floating_ips return a count
This modifies delete_unattached_floating_ips to return either a count
of how many floating IPs were cleaned, or False if none were cleaned.

Although this is an API change, from searching both codesearch and
google it seems very likely that nodepool only user of this.  However,
since a positive integer value will evaluate True anyway extant code
should continue to work.

The neturon test-case is updated with an extra floating-ip to make
sure the count works.  The nova test-case is updated to ensure it
returns False to indicate nothing was done.

Sort-of-Needed-By: https://review.openstack.org/614074
Change-Id: I7bd709bd83f352c58203c767779acbe66ecfc10e