Filter ports in list_ports when batching is in effect
When caching is in effect, we always fetch ports in a raw list with no
push-down filters. However, people calling list_ports with the filters
argument should expect it to work whether push-down conditions are used
or not. Wrap the return in a filter list so that we apply the filters
client-side in that case.

Story: 2004207
Task: 27716
Change-Id: Ieabb3193c320bd9c0e569e2e7cef983529a6fbdb