Document "insecure" flag
I received a clouds.yaml with "insecure: True" which I had not seen
before.

There's no explicit mention of this in the documentation, but due to a
name clash in the example in the reference it initially led me to
believe it was a section of the config file.  So firstly make the
cloud names more verbose to illustrate what's going on, and add a
mention of the "insecure" flag, but suggest using "verify".

Similarly in the unit tests use more explicit cloud names to make it
obvious what's going on.  Add a test that ensures "insecure=True"
implies "verify=False".

Change-Id: I56fe7460431442ad4ca264e32d2494068cf7f05e