Support remote vendor profiles
Maintaining vendor profile information inside of the sdk is great and
all, but it winds up being problematic from a release management
perspective since the data should always be current but people may not
be in a position to upgrade their version of openstacksdk. It's also a
less pleasing experience for people running or using clouds that the
openstacksdk developers don't know about.

RFC 5785 defines a scheme for serving data at well known URL locations.
Use it to allow specifying a profile by URL instead of by name. For
instance, for the cloud Example, a user could list profile:
https://example.com and openstacksdk will fetch the profile from
https://example.com/.well-known/openstack/api. It should be noted that
sub-urls are not allowed, so it MUST be served off of a root domain.
(That is, https://example.com/cloud is not allowed by the RFC)

Clouds are not required to serve one of these.

Change-Id: I884f62b35da5f29aa6e72e2dde9b8ec3ef48ad60