Register proxy directly in add_service
When adding an unknown service to Connection, the existing code was
adding the service's ServiceDescription. Unfortunately, adding a
Descriptor at runtime doesn't actually work and the added instance winds
up being unusable.

As add_service is an active call, we can assume that the person calling
it actually wants to use the service, so we don't need to hid it behind
an on-demand descriptor. Run _make_proxy directly and attach the
resulting proxy to the Connection.

Change-Id: I02aa1ebbd11bea33712a3cb82600abc1aed2a65a