Check result of server metadata operation
An accidental failure in one of the functional tests showed, that
response of metadata change was not checked. In that particular example
nova returned 500 for one of the metadata deletes, but the testcase
has not recognized this and went further, what resulted in a later
asserEquals failure.
So what we need to do here is to check the response of the metadata
operation and raise an exception if it was not successful.

Change-Id: I37bc96531eb37f9f850d5531354bd4db17973406