Avoid dogpile.cache 0.7.0
Latest release of dogpile.cache breaks our method of calling a bound
function in the decorators.  For more details see story or

  https://github.com/sqlalchemy/dogpile.cache/issues/140

Pin while we work on a solution.

Additionally, the cirros version was updated to 0.3.6 in devsatck with
Id2f20ebafdd78c2dadf81b8f80f22e7bd6db7755.  Update the hard-coded
versions in the functional tests (this should be modified to be more
version agnostic in a future change)

Story: 2004605
Depends-On: https://review.openstack.org/624697
Change-Id: Id2df8a05ef45041bf55a2671f7d06e7b88cc79d6