Properly munch for resource sub-dicts
In the shade layer, we expect object notation to work for sub-dicts.
When we're using underlying resource objects and translating them to
munch then putting them through normalize (Which is temporary during
transition) we're losing the munchified sub-dicts.

Update to_dict in openstack/resource to be able to provide munches
instead of dicts so that the recursive transform is complete.

Add a test for server that makes sure we're getting what we need.

A followup patch that should come that sets original_names to false in
the to_munch call, which will need an update to the normalize function
to deal with new incoming name.

Change-Id: I3df806fe0db7ddf8d93546d64780fc979f38e78f