Start using direct REST in normalize tests
This is too synthetic and we're missing things. Switch to requests_mock.

In doing this, remove the extra creation of extra connection objects in
the base test. Reorganize the normalize test file so that they just drop
in a strict_cloud flag and get the cloud created properly the first
time.

Change-Id: I86436b439cbc4b236b806af7eed195b198ce54e8