Compute location properly in server
The shade layer pulls the az and project_id info out of the server
record and feeds it into the location field. This is because the
location field should be where the resource is, and the existing token
should only be a fallback.

This isn't happening in the resource layer, so start doing that.

As a result of digging, make sure we always pass in the connection
object to the Resources when we make them. And stop creating a bare
resource in order to use the 'new' class method. It's a class method, we
don't need to instantiate.

Change-Id: Ic8850cd2863e2c806464b92cde2c749f6cc01a91