Add possibility to override base_path for resource operations
We currently have multiple places, where we inherit resource just to be
able to request it's Details or so under different url. With 619594
a change has been introduced to improve this by having possibility
to override base_path for a single resource operation.
This was done only for list operations. However there are also cases
when such possibility would be nice for create/update/etc operations.
One example is dry-run for heat, where the endpoint receives "/preview"
part. This is currently implemented by creating
additional resource StackPreview with different base_path. For create
it might be ok, but dry-run stack update would require another url.

With this change a possibility to override base_path for individual
operation on resource object is added. It can be given to proxy
function, which would pass it to resource and use it during URI
calculation.

Change-Id: Id8c8212249cb985d2e47eb1d4fb23ebf19b3871b