Skip v2 block-storage tests when service is not found
We have skips in for the cloud functional tests. Add the same skips for the
resource layer tests.

Since the resource layer is tied to version 2, put in an explicit
config override in the setUp to set block_storage_api_version.

Shift block_storage v2 tests to use self.user_cloud and not self.conn.

In doing so, test_type turns out to need admin privs, so use self.op_cloud
for it.

Change-Id: I8b231e7b5e0ebad6751702d9f0cb85e3d2192ac5