Make all resource locations process project_id
All resources have a location property now that describes the
location of the object in the cloud, but the majority of that
information is taken directly from the connection that is in
use. Some resources contain server-side project_id or availability
zone information. When those exist, they should override the
project_id implied by the connection.

Change-Id: I48c116bf62d726be8d906c3d232e1ee203347b0a