Use mock context in test_fwaas
Instead of mocking objects and restoring the original value at the
end of the test, this patch uses the mock context.

The original implementation is prone to errors. If the test doesn't
finish correctly, the original object is never restored. However inside
a mock context, if the test exits prematurely, the original object is
restored.

Change-Id: Ia333230b906470e27da326b7365049f79b7dcf6a