Tweak find_image method to search in hidden images
With Image v2.7 a possibility of hiding and image is being added (an 
attribute 'os_hidden'). After an image is being hidden the only 
possibility to find image by name is to do a list with `os_hidden=True` 
to list hidden images. So what we do here is search as usual, and if 
nothing found try also to search in hidden images (list of hidden images 
and grep by name)

Change-Id: I237da3e7e6b5d23010d8b26a3fcbfabaddeb5f31