Move object methods to object_store proxy
This is a similar change to https://review.openstack.org/#/c/609684/ 
with the purpose to streamline the code between SDK/shade and give 
ability OSC to use the well tested logic.

NOTE: This is not yet finished. The proxy methods should be transformed
to using the Resource layer. This requires enabling custom meta headers 
of the object to avoid doing direct PUT and GET bypassing the resource 
(also doing too much endpoint calculation)

Change-Id: I5d288c25797833a850b361af8f067923b00b8706