Swap human-facing links to use opendev.org
While we should not start using this for cloning things in the gate yet,
the opendev.org links provide for a nicer browsing experience already
even though the service is in beta. Go ahead and swap our browsing
links.

While in there, remove some masakari references to github links to the
source code. Masakari now publishes api docs to developer.openstack.org
so there is no need to point people to source code for api docs.

Change-Id: I14afa3476f9832ef433ec744e888348fc9a359e3